Create functions for converting Prio values
Added a selector in styles.css to target the input
that has type="text" with targeting the ID.

The functions in converting works in simple way.
2 functions each specialize in converting either
from number to text or text to number (1 - High etc).
The 2 function are used in a one function depending
on their return value - hence the logical operator.