Attach "submit" event to form itself instead of button
The button handled the e.preventDefault for form
submission. Bet when "required" was added to some
inputs it didnt work -> the button closes the
form by defined methods - the form isnt submitted
at all so required fields are not really required.

When change the event from click to submit it still
didnt work - the required fields worked, but the
form was reloaded with POST - the button doesnt really
fire the "submit" event - the form does - so adding
the event listener with adequate event to form
solves this issue