Add and Load created Projects from localStorage
the localStorage has 1 main item from which
everything creates "projects". Project Menu loads
the sidebar with created Projects - it is never
empty on reload (always a default created project
for the user).
- if no projects gotten from localStorage then
  create a default one
- if there are then load them and display on the
  sidebar

3 main occasions - first load, delete and add
in every event the item of localStorage is set to
the latest projectManager.projects array.

Also changed the position of parameters in project.js
so the right parameters are given.